Offloaded the cooldown cast breakdowns to their respective spell modules, and made CooldownExpandable component a little less try hard
