--- analysis/druidrestoration/src/modules/features/Tranquility.ts
@@ -1,80 +1,0 @@
-import SPELLS from 'common/SPELLS';
-import Analyzer, { Options, SELECTED_PLAYER } from 'parser/core/Analyzer';
-import Events, { CastEvent, HealEvent } from 'parser/core/Events';
-
-import { getTranquilityTicks } from '../../normalizers/CastLinkNormalizer';
-import HotTrackerRestoDruid from '../core/hottracking/HotTrackerRestoDruid';
-
-export const MAX_TRANQ_TICKS = 5;
-
-/**
- * Tracks stats relating to Wild Growth
- */
-class Tranquility extends Analyzer {
-  static dependencies = {
-    hotTracker: HotTrackerRestoDruid,
-  };
-
-  hotTracker!: HotTrackerRestoDruid;
-
-  tranqCasts: TranquilityCast[] = [];
-
-  constructor(options: Options) {
-    super(options);
-    this.addEventListener(
-      Events.cast.by(SELECTED_PLAYER).spell(SPELLS.TRANQUILITY_CAST),
-      this.onTranqCast,
-    );
-    this.addEventListener(
-      Events.heal.by(SELECTED_PLAYER).spell(SPELLS.TRANQUILITY_HEAL),
-      this.onTranqHeal,
-    );
-  }
-
-  onTranqCast(event: CastEvent) {
-    const directHealing = 0;
-    const periodicHealing = 0;
-    const rejuvsOnCast =
-      this.hotTracker.getHotCount(SPELLS.REJUVENATION.id) +
-      this.hotTracker.getHotCount(SPELLS.REJUVENATION_GERMINATION.id);
-    const wgsOnCast = this.hotTracker.getHotCount(SPELLS.WILD_GROWTH.id);
-    const timestamp = event.timestamp;
-    const channeledTicks = getTranquilityTicks(event).length;
-    this.tranqCasts.push({
-      timestamp,
-      directHealing,
-      periodicHealing,
-      wgsOnCast,
-      rejuvsOnCast,
-      channeledTicks,
-    });
-  }
-
-  onTranqHeal(event: HealEvent) {
-    const effectiveAmount = event.amount + (event.absorbed || 0);
-    if (this.tranqCasts.length > 0) {
-      if (event.tick) {
-        this.tranqCasts[this.tranqCasts.length - 1].periodicHealing += effectiveAmount;
-      } else {
-        this.tranqCasts[this.tranqCasts.length - 1].directHealing += effectiveAmount;
-      }
-    }
-  }
-}
-
-interface TranquilityCast {
-  /** Timestamp of the start of the Tranquility channel */
-  timestamp: number;
-  /** The healing from this cast's direct portion */
-  directHealing: number;
-  /** The healing from this cast's HoTs */
-  periodicHealing: number;
-  /** The number of Wild Growths out at the moment this Convoke is cast */
-  wgsOnCast: number;
-  /** The number of Rejuvs out at the moment this Convoke is cast */
-  rejuvsOnCast: number;
-  /** The number of ticks that were channeled in this cast */
-  channeledTicks: number;
-}
-
-export default Tranquility;
